skills/video_summary/pyproject.toml
[project]
name = "video_summary"
version = "0.1.0"
description = "Video transcription and summary generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "common",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_summary"]

[tool.uv.sources]
common = { workspace = true }
